==> notebooks/06_redindex_amazon.ipynb <==
# Preparação do ambiente

## Bibliotecas

import pandas as pd
from os import path as osp

## Constantes e funções auxiliares

NULL_STRING = '#NULLO#'

# Carga e prepação dos dados

## Reindexar

Reindexar dados da Amazon. Os dados reindexados serão usados como base para o desenvolvimento de experimentos de classificação.

file_amazon_raw = '../datasets/amazon/raw/amazon.parquet.gzip'
file_amazon_reindexed = '../datasets/amazon/raw/amazon_reindexed.parquet'

if osp.exists(file_amazon_reindexed):
    print(f'Read existing file {file_amazon_reindexed}.')
    df_amazon = pd.read_parquet(file_amazon_reindexed)
    
else:
    print(f'Read existing file {file_amazon_raw}, redindex and save to new file {file_amazon_reindexed}')
    df_amazon = pd.read_parquet(file_amazon_raw)
    # fill na values with NULL_STRING
    df_amazon = df_amazon.fillna(NULL_STRING)
    # reindex
    df_amazon = df_amazon.reset_index(drop=True)
    # save to new file
    df_amazon.to_parquet(file_amazon_reindexed,compression='gzip')
df_amazon

columns_to_keep = ['cat_0','cat_1','cat_2','cat_3','cat_4','cat_5']
df_amazon['cat_full']=df_amazon[columns_to_keep].apply(lambda row: '_'.join(row),axis=1)
df_amazon['cat_full'].value_counts()

df_cat_count = df_amazon.cat_0.value_counts()
df_cat_count

file_cat_to_label = '../datasets/amazon/targets/amazon_cat_to_label.xlsx'

columns_to_keep = ['cat_0','cat_1','cat_2','cat_3','cat_4','cat_5','cat_full','Nome']
df_cat = df_amazon[df_amazon['cat_0']!=NULL_STRING][columns_to_keep]
df_cat = df_cat.groupby(columns_to_keep[:-1],as_index=False).count().sort_values(by='Nome',ascending=False)

# exportar em xlsx para manipular os dados manualmente
# caso já exista, exportar em formato csv para evitar perda dos dados já manipulados
if osp.exists(file_cat_to_label):
    file_cat_to_label_csv = file_cat_to_label.replace('xlsx', 'csv')
    print(f'Alert! File {file_cat_to_label} exists. Save to file{file_cat_to_label_csv}')
    df_cat.to_csv(file_cat_to_label_csv,index=False,sep=';')
else:    
    print(f'Save to file{file_cat_to_label}')
    df_cat.to_excel(file_cat_to_label,index=False)
df_cat

columns_to_keep = columns_to_keep = ['cat_0','cat_1','cat_2','cat_3','cat_4','cat_5','cat_full','Link','Nome']
df_cat_sample = df_amazon[columns_to_keep]

# map cat with less than 1000 items
map_cat_outros = {cat:'Outros' for cat in df_cat_count[df_cat_count<1000].index}
df_cat_sample.loc[:,'cat_full'] = df_cat_sample[['cat_0','cat_full']].apply(lambda cat: map_cat_outros[cat['cat_0']] if cat['cat_0'] in map_cat_outros.keys() else cat['cat_full'],axis=1)

df_cat_sample = df_cat_sample.groupby('cat_full').sample(10,replace=True,random_state=22000).drop_duplicates()

if osp.exists(file_cat_to_label):
    file_cat_to_label_csv = file_cat_to_label.replace('xlsx','csv')
    print(f'Alert! File {file_cat_to_label} exists. Save to file {file_cat_to_label_csv}')
    df_cat_sample.to_csv(file_cat_to_label_csv,index=False,sep=';')
else:    
    print(f'Save to file {file_cat_to_label}')
    df_cat_sample.to_excel(file_cat_to_label,index=False)
df_cat_sample

# Outros experimentos

# df_actual_sample = pd.read_excel(file_cat_sample,sheet_name='CategoriasTargets')

# df_amazon_actual_sample_indexes = df_amazon[['Link', 'Imagem', 'Imagens']].reset_index()
# df_amazon_actual_sample_indexes = df_amazon_actual_sample_indexes.merge(df_actual_sample['Link'])

# df_amazon_actual_sample_indexes['Imagem'] = df_amazon_actual_sample_indexes['Imagem'].apply(lambda x: x.split(' ')[0])
# df_amazon_actual_sample_indexes['Imagens'] = df_amazon_actual_sample_indexes['Imagens'].apply(lambda x: x.replace('\n',','))

# file_amazon_actual_sample_indexes = '../datasets/actual_sample_link_images.xlsx'
# df_amazon_actual_sample_indexes.to_excel(file_amazon_actual_sample_indexes)

# from nltk.tokenize import word_tokenize
# from nltk.corpus import stopwords
# import string

# stop_words = stopwords.words('portuguese')
# stop_words.extend(stopwords.words('english'))
# stop_words.extend(list(string.punctuation))

# words = word_tokenize('Mingzhe Transmissor de controle remoto FS-CT6B com receptor . FS-R6B 2.4G. 6CH para RC carro barco quadricóptero planador helicóptero')
# print(words)
# wc = [w for w in words if w not in stop_words]
# print(wc)
