==> notebooks/17_new_classifier_testing.ipynb <==
# Preparação do ambiente

## Bibliotecas

import json
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd

from itertools import pairwise
from pickle import load

## Constantes e funções auxiliares

FILE_SCH_DATABASE = '../datasets/sch/produtos_certificados.zip'
FILES_INSPECAO_ECOMMERCE = [
    '../datasets/inspecao_ecommerce/carrefour.parquet.gzip',   
    '../datasets/inspecao_ecommerce/ml.parquet.gzip', 
    '../datasets/inspecao_ecommerce/magalu.parquet.gzip']
FILE_MODEL = '../datasets/model/sgd_model_20240423.pkl'

def load_sch(sch_database_file):
    
    # load SCH database
    usecols = [0,1,11,12,13,14,15]
    dtype = {'Número de Homologação': 'str'}
    parse_dates = [0]
    date_format = '%d/%m/%Y'

    df_sch = pd.read_csv(
        sch_database_file,
        sep=';',
        usecols=usecols,
        dtype=dtype,
        parse_dates=parse_dates,
        date_format=date_format
        )

    df_sch = df_sch.sort_values(by='Data da Homologação',ascending=False)

    df_modelo = df_sch[['Número de Homologação','Modelo']].dropna()
    df_modelo = df_modelo.groupby('Número de Homologação',as_index=False)['Modelo'].apply(lambda x: ' | '.join(x))

    df_nome_comercial = df_sch[['Número de Homologação','Nome Comercial']].dropna()
    df_nome_comercial = df_nome_comercial.groupby('Número de Homologação',as_index=False)['Nome Comercial'].apply(lambda x: ' | '.join(x))

    columns_to_keep = ['Data da Homologação', 'Número de Homologação', 'Nome do Fabricante', 'Categoria do Produto', 'Tipo do Produto']
    df_sch = df_sch[columns_to_keep]
    df_sch = df_sch.drop_duplicates(subset='Número de Homologação')

    df_sch = df_sch.merge(df_modelo,how='left')
    df_sch = df_sch.merge(df_nome_comercial,how='left')
    df_sch = df_sch.fillna('')
    
    df_sch = df_sch.reset_index(drop=True)
    
    return df_sch 

def flatten_matrix(matrix):
    return [item for row in matrix for item in row]

def extract_amazon_images(row):
    imagens = []
    imagens.append(row['Imagem'].split()[0])
    imagens.extend(row['Imagens'].split())

    imagens = [imagem for imagem in imagens if imagem.startswith('http')]
    if len(imagens) >= 3:
        imagens = imagens[:3]
    elif len(imagens) == 2:
        imagens.extend([''])
    elif len(imagens) == 1:
        imagens.extend(['',''])
    else:
        imagens.extend(['','',''])

    return pd.Series(imagens)

def extract_amazon_desc(row):
    desc = '\n'.join(row)
    desc = desc.replace(NULL_STRING,' ')
    desc = re.sub('\s+', ' ', desc)
    return desc

def extract_images(row):
    imagens = []
    imagens.append(row['Imagem'].split()[0])
    
    if not pd.isna(row['Imagens']):
        imagens.extend(row['Imagens'].split(','))

    imagens = [imagem for imagem in imagens if imagem.startswith('http')]
    if len(imagens) >= 3:
        imagens = imagens[:3]
    elif len(imagens) == 2:
        imagens.extend([''])
    elif len(imagens) == 1:
        imagens.extend(['',''])
    else:
        imagens.extend(['','',''])

    return pd.Series(imagens)

def format_issue(row,remove_punct=False):
    nome,desc,img0,img1,img2 = row
    
    # remove null string
    desc = desc.replace(NULL_STRING,'')
    # remove punctiation from nome and desc
    if remove_punct:
        nome = nome.translate(str.maketrans('', '', string.punctuation))
        desc = desc.translate(str.maketrans('', '', string.punctuation))

    # remove ponto-e-vírgula e aspas duplas para evitar a 
    # identificação incorreta de campos ao importar para o redmine
    desc = desc.replace(';',',')
    desc = desc.replace('"',"''")
    # remove multiple linebrakes and spaces
    desc = re.sub('\n+','\n\n',desc)
    desc = re.sub('\s+',' ',desc)
    desc = desc.strip()
    
    img_css = '{max-width:250px; height:auto}'
    if len(img2) > 0:
        img_list = f'!{img_css}{img0}! !{img_css}{img1}! !{img_css}{img2}!'
    elif len(img1) > 0:
        img_list = f'!{img_css}{img0}! !{img_css}{img1}!'
    elif len(img0) > 0:
        img_list = f'!{img_css}{img0}!'
    else:
        img_list = ''
        
    if len(desc) >= 1:
        issue = f'h1. {nome}\n\n{desc}\n\n{img_list}'
    else:
        issue = f'h1. {nome}\n\n{img_list}'
    
    return issue.strip()

def plot_donut(labels,values,title,startangle=30):
    fig, ax = plt.subplots(figsize=(5, 3), subplot_kw=dict(aspect="equal"))
    
    wedges, texts = ax.pie(values, wedgeprops=dict(width=0.5), startangle=startangle)
    
    bbox_props = dict(boxstyle="round,pad=0.3", fc="w", ec="k", lw=0.72)
    kw = dict(arrowprops=dict(arrowstyle="-"),
              bbox=bbox_props, zorder=0, va="center")
    
    kw = dict(arrowprops=dict(arrowstyle="-"),bbox=bbox_props,
              zorder=0, va="center")
    
    for i, p in enumerate(wedges):
        ang = (p.theta2 - p.theta1)/2. + p.theta1
        y = np.sin(np.deg2rad(ang))
        x = np.cos(np.deg2rad(ang))
        horizontalalignment = {-1: "right", 1: "left"}[int(np.sign(x))]
        connectionstyle = "angle,angleA=0,angleB={}".format(ang)
        kw["arrowprops"].update({"connectionstyle": connectionstyle})
        ax.annotate(labels[i], xy=(x, y), xytext=(1.35*np.sign(x), 1.4*y),
                    horizontalalignment=horizontalalignment, **kw)
    
    ax.set_title(title)
    plt.show()

# Carga e prepação dos dados

## Base SCH

df_sch = load_sch(FILE_SCH_DATABASE)
df_sch.head()

## Anúncios coletados

df_inspecao_ecommerce = pd.concat([pd.read_parquet(file) for file in FILES_INSPECAO_ECOMMERCE])
df_inspecao_ecommerce = df_inspecao_ecommerce.reset_index(drop=True)
df_inspecao_ecommerce['Descrição'] = df_inspecao_ecommerce['Descrição'].fillna('')
df_inspecao_ecommerce

## Modelo pré-treinado

with open(FILE_MODEL, 'rb') as f:
    clf = load(f)
clf

# Análise

# df.iloc[5013]
# df_sch[df_sch['Número de Homologação']=='076211902585']

docs = df_inspecao_ecommerce['Nome'].values
passivel = clf.predict(docs)
proba = clf.predict_proba(docs)[:,1]

map_passivel = {0: 'Não', 1: 'Sim'}

df_inspecao_ecommerce['Passível'] = passivel
df_inspecao_ecommerce['Passível'] = df_inspecao_ecommerce['Passível'].map(map_passivel)
df_inspecao_ecommerce['Probabilidade'] = proba

df_inspecao_ecommerce.head()

df_resumo = df_inspecao_ecommerce.pivot_table(index='Palavra_Chave',columns='Passível', values='Nome', aggfunc='count').fillna(0)
df_resumo = df_resumo.astype('int')
df_resumo['Total'] = df_resumo.sum(axis=1)
df_resumo.loc['Total'] = df_resumo.sum()
df_resumo
