==> notebooks/18.1_recover_issues.ipynb <==
# Preparação do ambiente

## Bibliotecas

import pandas as pd
import re
import warnings

from datetime import datetime
from datetime import timedelta

from pathlib import Path

pd.options.mode.chained_assignment = None  # default='warn'
# Suppress FutureWarning messages
warnings.simplefilter(action='ignore', category=FutureWarning)

## Constantes e funções auxiliares

ISSUES_FOLDER = Path(r'../datasets/inspecao_ecommerce/issues')
FILE_CLOSED_ISSUES = '../datasets/inspecao_ecommerce/avaliacao_regulatron/closed_issues.csv'
FILE_ANUNCIOS_REGULATRON = '../datasets/inspecao_ecommerce/avaliacao_regulatron/Anuncios.xlsx'

# Carga e prepação dos dados

## Tarefas de rotulagem (segunda fiscalização)

files_issues = [file for file in ISSUES_FOLDER.rglob('*.csv')]

df_issues = pd.concat([pd.read_csv(file,sep=';',quotechar="'") for file in files_issues])
df_issues['Qual foi a data da coleta?'] = pd.to_datetime(df_issues['Qual foi a data da coleta?'],format='mixed')
df_issues['Qual foi a data da coleta?'] = df_issues['Qual foi a data da coleta?'].dt.date
df_issues['Arquivo PDF'] = df_issues['Descrição'].apply(
    lambda descricao: re.search('https.+pdf',descricao).group() 
    if re.search('https.+pdf',descricao) 
    else None
)
df_issues['nome'] = df_issues['Descrição'].apply(lambda descricao: descricao[3:].split('\n')[0])

df_issues.head()

## Resultados da segunda fiscalização


df_resultados = pd.read_csv(FILE_CLOSED_ISSUES,sep=';')
df_resultados = df_resultados[
    (df_resultados['Marketplace']=='Amazon')
    &(df_resultados['Texto da busca']=='smartphone')
]

columns_to_keep = ['Título', 'url', 'Arquivo PDF']
# df_issues[df_issues['Marketplace']=='Amazon'][columns_to_keep].to_excel('../datasets/inspecao_ecommerce/avaliacao_regulatron/issues_screenshots.xlsx')
df_resultados = df_resultados.merge(
    df_issues[df_issues['Marketplace']=='Amazon'][columns_to_keep]
)
# estabelecer grupo por período de coleta
df_resultados['coleta'] = '2024-07-08'
df_resultados['nome'] = ''

columns_to_keep = ['url', 'Arquivo PDF', 'coleta', 'nome', 'Qual o tipo do produto? ', 'Fabricante', 'Modelo do Produto', 'Qual código SCH fornecido?', 'Qual código EAN fornecido?']
df_resultados = df_resultados[columns_to_keep]

new_columns = ['url', 'screenshot', 'coleta', 'nome', 'tipo_subcategoria', 'fabricante', 'modelo', 'certificado', 'ean_gtin']
df_resultados.columns = new_columns

df_resultados

## Raspagem do Regulatron

%%time
dtype = {
    'certificado': 'str',
    'ean_gtin': 'str'
}

# importar resultados de raspagem
df_anuncios_regulatron = pd.read_excel(FILE_ANUNCIOS_REGULATRON,dtype=dtype)

# manter apenas Amazon
df_anuncios_regulatron = df_anuncios_regulatron[df_anuncios_regulatron['marketplace']=='Amazon']
# manter apenas busca por smartphone
df_anuncios_regulatron = df_anuncios_regulatron[df_anuncios_regulatron['palavra_busca'].str.contains('smartphone')] 

df_anuncios_regulatron['certificado'] = df_anuncios_regulatron['certificado'].fillna('0')
df_anuncios_regulatron['ean_gtin'] = df_anuncios_regulatron['ean_gtin'].fillna('0')

# agrupar coletas
map_coletas = {
    '2024-10-02': '2024-10-02', 
    '2024-10-03': '2024-10-02',
    '2024-10-04': '2024-10-02',
    '2024-10-10': '2024-10-10',
    '2024-10-11': '2024-10-10',
    '2024-10-22': '2024-10-22',
}
df_anuncios_regulatron['coleta'] = df_anuncios_regulatron.data.apply(lambda data: data[:10]).map(map_coletas)


# # manter apenas urls que estavam na primeira fiscalização (julho)
# df_anuncios_regulatron = df_anuncios_regulatron.merge(df_espatula.url)
# # contar quantas vezes cada url aparece
# df_anuncios_regulatron['url_count'] = df_anuncios_regulatron[['url','nome']].groupby('url').transform('count')

columns_to_keep = ['url', 'screenshot', 'coleta', 'nome', 'subcategoria', 'fabricante', 'modelo', 'certificado', 'ean_gtin']
df_anuncios_regulatron = df_anuncios_regulatron[columns_to_keep]

df_anuncios_regulatron.columns = new_columns
df_anuncios_regulatron

# Anúncios comuns

# unir os dataframes
df_anuncios_analise = pd.concat([df_resultados, df_anuncios_regulatron])
df_anuncios_analise.coleta = pd.to_datetime(df_anuncios_analise.coleta)
df_anuncios_analise.modelo = df_anuncios_analise.modelo.fillna('0')

sort_keys = ['url', 'coleta']
df_anuncios_analise = df_anuncios_analise.sort_values(by=sort_keys)
df_anuncios_analise = df_anuncios_analise.drop_duplicates(subset=sort_keys)

df_anuncios_analise['certificado_anuncio'] = df_anuncios_analise['certificado']!='0'
df_anuncios_analise['certificado_anuncio'] = df_anuncios_analise['certificado_anuncio'].astype(int)

df_anuncios_analise['ean_gtin_anuncio'] = df_anuncios_analise['ean_gtin']!='0'
df_anuncios_analise['ean_gtin_anuncio'] = df_anuncios_analise['ean_gtin_anuncio'].astype(int)

# df_anuncios_analise = df_anuncios_analise.drop_duplicates(subset='url')
# df_anuncios_analise = df_anuncios_analise.reset_index(drop=True)
# df_anuncios_analise.to_excel('../datasets/inspecao_ecommerce/avaliacao_regulatron/anuncios_analise_historico.xlsx')

df_anuncios_analise['ocorrencias'] = df_anuncios_analise[['url','modelo']].groupby('url').transform('count')
# df_anuncios_analise = df_anuncios_analise[df_anuncios_analise['ocorrencias']>2]

df_anuncios_analise = df_anuncios_analise.reset_index(drop=True)
df_anuncios_analise.to_excel('../datasets/inspecao_ecommerce/avaliacao_regulatron/anuncios_para_limpar_2.xlsx')
df_anuncios_analise

df_anuncios_limpos = pd.read_excel('../datasets/inspecao_ecommerce/avaliacao_regulatron/anuncios_para_limpar.xlsx')
df_anuncios_limpos

df_resumo_todos = df_anuncios_limpos.groupby('coleta').agg({'url': 'count', 'certificado_anuncio': 'sum', 'ean_gtin_anuncio':'sum'})
df_resumo_todos['%_certificado'] = df_resumo_todos['certificado_anuncio']/df_resumo_todos['url']
df_resumo_todos['%_ean_gtin'] = df_resumo_todos['ean_gtin_anuncio']/df_resumo_todos['url']

df_resumo_todos

df_resumo_mais_que_2 = df_anuncios_limpos[df_anuncios_limpos['ocorrencias']>2].groupby('coleta').agg({'url': 'count', 'certificado_anuncio': 'sum', 'ean_gtin_anuncio':'sum'})
df_resumo_mais_que_2['%_certificado'] = df_resumo_mais_que_2['certificado_anuncio']/df_resumo_mais_que_2['url']
df_resumo_mais_que_2['%_ean_gtin'] = df_resumo_mais_que_2['ean_gtin_anuncio']/df_resumo_mais_que_2['url']

df_resumo_mais_que_2

df_resumo_4 = df_anuncios_limpos[df_anuncios_limpos['ocorrencias']==4].groupby('coleta').agg({'url': 'count', 'certificado_anuncio': 'sum', 'ean_gtin_anuncio':'sum'})
df_resumo_4['%_certificado'] = df_resumo_4['certificado_anuncio']/df_resumo_4['url']
df_resumo_4['%_ean_gtin'] = df_resumo_4['ean_gtin_anuncio']/df_resumo_4['url']

df_resumo_4

df_anuncios_regulatron.pivot_table(index='url_count', columns='data', values='nome', aggfunc='count').fillna(0).astype(int)
