==> notebooks/19_poc_api_meli.ipynb <==
# Preparação do ambiente

## Bibliotecas

import json
import pandas as pd
import requests
import urllib.parse

from pickle import load
from tqdm.auto import tqdm

# import sys
# import codecs
# import urllib.request

pd.options.mode.chained_assignment = None  # default='warn'

## Constantes e funções auxiliares

file_model = '../models/clf_market_places_sgd.pkl'
with open(file_model, 'rb') as f:
    clf = load(f)
clf

USER_AGENT = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/126.0.0.0 Safari/537.36'

def load_sch(sch_database_file):
    
    # load SCH database
    usecols = [0,1,11,12,13,14,15]
    dtype = {'Número de Homologação': 'str'}
    parse_dates = [0]
    date_format = '%d/%m/%Y'

    df_sch = pd.read_csv(
        sch_database_file,
        sep=';',
        usecols=usecols,
        dtype=dtype,
        parse_dates=parse_dates,
        date_format=date_format
        )

    df_sch = df_sch.sort_values(by='Data da Homologação',ascending=False)

    df_modelo = df_sch[['Número de Homologação','Modelo']].dropna()
    df_modelo = df_modelo.groupby('Número de Homologação',as_index=False)['Modelo'].apply(lambda x: ' | '.join(x))

    df_nome_comercial = df_sch[['Número de Homologação','Nome Comercial']].dropna()
    df_nome_comercial = df_nome_comercial.groupby('Número de Homologação',as_index=False)['Nome Comercial'].apply(lambda x: ' | '.join(x))

    columns_to_keep = ['Data da Homologação', 'Número de Homologação', 'Nome do Fabricante', 'Categoria do Produto', 'Tipo do Produto']
    df_sch = df_sch[columns_to_keep]
    df_sch = df_sch.drop_duplicates(subset='Número de Homologação')

    df_sch = df_sch.merge(df_modelo,how='left')
    df_sch = df_sch.merge(df_nome_comercial,how='left')
    df_sch = df_sch.fillna('')
    df_sch['Modelo Completo'] = df_sch[['Modelo', 'Nome Comercial']].apply(lambda row: ' | '.join(row) if len(row['Nome Comercial'])>0 else row['Modelo'],axis=1)
    
    df_sch = df_sch.reset_index(drop=True)
    
    return df_sch 

def parse_results(result, query=None):
    keys_to_keep = ['id', 'title', 'catalog_product_id', 'permalink', 'category_id', 'domain_id', 
                    'currency_id', 'price', 'original_price', 'sale_price', 
                    'initial_quantity', 'available_quantity', 'official_store_id', 'official_store_name']
    
    parsed_result = {}
    parsed_result['query'] = query
    for key in keys_to_keep:
        parsed_result[key] = result.get(key)
    
    if result.get('seller') is not None:
        parsed_result['seller_id'] = result.get('seller').get('id')
        parsed_result['seller_nickname'] = result.get('seller').get('nickname')
    else:
        parsed_result['seller_id'] = None
        parsed_result['seller_nickname'] = None    

    brand = None
    model = None
    gtin = None
    detailed_model = None
    anatel_homologation_number = None
    
    for attribute in result['attributes']:
        if attribute['id'] == 'BRAND':
            brand = attribute['value_name']
        elif attribute['id'] == 'MODEL':
            model = attribute['value_name']
        elif attribute['id'] == 'DETAILED_MODEL':
            detailed_model = attribute['value_name'] 
        elif attribute['id'] == 'GTIN':
            gtin = attribute['value_name'] 
        elif attribute['id'] == 'ANATEL_HOMOLOGATION_NUMBER':
            anatel_homologation_number = attribute['value_name']

    if anatel_homologation_number is None:
        item_id = parsed_result['id']
        attr_url = f'https://api.mercadolibre.com/items/{item_id}?attributes=attributes&include_internal_attributes=true'
        headers = {'user-agent': USER_AGENT}
        attr_response = requests.get(attr_url,headers=headers)
        attr_content = attr_response.content.decode(attr_response.encoding)
        attr_content =  json.loads(attr_content)
        attr_content = attr_content['attributes']
        for attr in attr_content:
            if attr['id'] == 'ANATEL_HOMOLOGATION_NUMBER':
                anatel_homologation_number = attr['value_name']

    parsed_result['brand'] = brand
    parsed_result['model'] = model
    parsed_result['detailed_model'] = detailed_model
    parsed_result['gtin'] = gtin
    parsed_result['anatel_homologation_number'] = anatel_homologation_number
    
    return parsed_result

keywords=['tv','box']
query = urllib.parse.quote_plus(' '.join(keywords))
url = 'https://api.mercadolibre.com/sites/MLB/search'
params={'q': query, 'offset': 0}
headers = {'user-agent': USER_AGENT}
response = requests.get(url,params=params)
response

def search_items(keywords):
    
    query = urllib.parse.quote_plus(' '.join(keywords))

    print('Start searching for', query)
    
    url = 'https://api.mercadolibre.com/sites/MLB/search'
    params={'q': query, 'offset': 0}
    headers = {'user-agent': USER_AGENT}
    
    results = []
    
    response = requests.get(url,params=params)
    content = response.content.decode(response.encoding)
    content =  json.loads(content)
    results.extend(content['results'])
        
    primary_results = content['paging']['primary_results']    
    # max offset shold be 1000, so we add 1 to primary_results
    next_offsets = list(range(50,primary_results+1,50))

    for offset in next_offsets:
        params={'q': query, 'offset': offset}
        response = requests.get(url,params=params)
        content = response.content.decode(response.encoding)
        content =  json.loads(content)
        results.extend(content['results'])

    print('Search complete: parsing results...')
    parsed_results = [parse_results(result,query) for result in tqdm(results)]
        
    return pd.DataFrame(parsed_results)

# Carga e prepação dos dados

FILE_SCH = '../datasets/sch/produtos_certificados.zip'
df_sch = load_sch(FILE_SCH)
df_sch.head()

FILE_EAN = '../datasets/sch/lista_celulares_homologados_ean.xlsx'

dtype = {'Número de Homologação': 'str', 'Código EAN': 'str'}

df_ean_celulares = pd.read_excel(FILE_EAN,dtype=dtype)
df_ean_celulares = df_ean_celulares[df_ean_celulares['Código EAN'].str.len()>=13]

columns_to_keep = ['Código EAN', 'Número de Homologação']
df_ean_sch = df_ean_celulares[columns_to_keep].drop_duplicates()
# df_ean_sch = df_ean_sch.dropna().reset_index(drop=True)
# df_ean_sch = df_ean_sch.astype('int64')

df_ean_sch.columns = ['ean_sch', 'sch_sch']
df_ean_sch.head()

FILE_API_ML = '../datasets/api_ml.parquet'

# query Mercado Livre
keywords = [['celular','smartphone'],['tv','box'],['carregador','celular']]

df_api_ml = pd.concat([search_items(keyword) for keyword in keywords])
df_api_ml.to_parquet(file_api_ml)

# Open previous query
df_api_ml = pd.read_parquet(file_api_ml)

df_api_ml.head()

# Análise

columns_to_keep = ['query', 'id', 'permalink', 'title','brand', 'model', 'gtin', 'anatel_homologation_number']
df_analise = df_api_ml[columns_to_keep]

columns_to_keep = ['query', 'id', 'permalink', 'title','brand', 'model', 'ean_mp', 'sch_mp']
df_analise.columns = columns_to_keep

# df_analise['ean_mp'] = df_analise['ean_mp'].fillna('0')
df_analise['ean_mp'] = df_analise['ean_mp'].str.split(',')
df_analise = df_analise.explode('ean_mp')
# df_analise['ean_ml'] = df_analise['ean_ml'].str.zfill(13)

df_analise['passivel'] = clf.predict(df_analise.title)
df_analise['passivel_proba'] = clf.predict_proba(df_analise.title)[:,1]

df_analise['ean_presente'] = df_analise['ean_mp'].apply(lambda x:  0 if x is None else 1)
df_analise['sch_presente'] = df_analise['sch_mp'].apply(lambda x:  0 if x is None else 1)

df_analise['sch_mp'] = df_analise['sch_mp'].fillna('0').str.zfill(12)

columns_to_merge = ['Número de Homologação', 'Tipo do Produto', 'Modelo Completo']
df_analise = df_analise.merge(df_sch[columns_to_merge], left_on='sch_mp', right_on='Número de Homologação',how='left')
df_analise['sch_valido'] = df_analise['Número de Homologação'].apply(lambda x:  0 if pd.isna(x) else 1)

# df_analise = df_analise.merge(df_ean_sch,left_on='ean_ml', right_on='ean_sch', how='left')
# df_analise['ean_ok'] = df_analise['ean_ml']==df_analise['ean_sch']
# df_analise['sch_ok'] = df_analise['sch_ml']==df_analise['sch_sch']

# columns_to_merge = ['Número de Homologação', 'Modelo Completo']
# df_analise = df_analise.merge(df_sch[columns_to_merge],left_on='sch_ml',right_on='Número de Homologação',how='left')

df_analise.head()

columns_to_group = ['query', 'id', 'passivel', 'ean_presente', 'sch_presente', 'sch_valido']
agg_func = {'id': 'count', 'passivel': 'sum', 'ean_presente': 'sum', 'sch_presente': 'sum', 'sch_valido': 'sum'}
df_resumo = df_analise[columns_to_group].groupby(columns_to_group[0]).agg(agg_func)
df_resumo['pe_ean_presente'] = df_resumo['ean_presente']/df_resumo['passivel']*100
df_resumo['pe_sch_presente'] = df_resumo['sch_presente']/df_resumo['passivel']*100
df_resumo['pe_sch_valido'] = df_resumo['sch_valido']/df_resumo['passivel']*100

columns_to_keep = ['Total de anúncios', 'Anúncios de Produtos Passíveis', 
                   'Anúncios com EAN Informado', 'Anúncios com código SCH Informado', 'Anúncios com código SCH Válido', 
                   '% Anúncios com EAN Informado', '% Anúncios com código SCH Informado', '% Anúncios com código SCH Válido']

df_resumo.columns = columns_to_keep
df_resumo.index.name = 'Palavra chave'

df_resumo
