==> notebooks/25_validation_persistent_model.ipynb <==
# Preparação do ambiente

## Bibliotecas

import numpy as np
import os.path as osp
import pandas as pd

from joblib import load

## Constantes e funções auxiliares

FILE_MODEL = '../models/clf_marketplaces.joblib'
FILE_SUPERVISAO_MERCADO = '../datasets/inspecao_ecommerce/supervisao_mercado.xlsx'

class MarketPlacesPredictor():

    def __init__(self, file_model):
        if osp.exists(file_model):
            self.model = load(file_model)
        else:
            self.model = None
            print(f'Model file "{file_model}" not found')

    def predict(self, X):
        if self.model is not None:
            predicted_class = self.model.predict(X)
            predicted_proba = self.model.predict_proba(X)[:,1]
            return np.column_stack((predicted_class,predicted_proba))
        else:
            print('Model is not loaded')
            return None


# Carga e prepação dos dados

dict_df_mercado = pd.read_excel(
    FILE_SUPERVISAO_MERCADO,
    sheet_name=None,
    usecols=[2,8,10],
    names=['texto_busca', 'descricao', 'passivel_homologacao'],
    true_values=['Sim', 'sim'],
    false_values=['Não','não'],
    na_values=['-'])

df_mercado = pd.concat(dict_df_mercado.values())
df_mercado = df_mercado.dropna()
df_mercado['passivel_homologacao'] = df_mercado['passivel_homologacao'].astype(int)
df_mercado

# Análise

predictor = MarketPlacesPredictor(FILE_MODEL)
predictor.model

X = df_mercado.descricao.values
y = df_mercado.passivel_homologacao.values

prediction = predictor.predict(X)

if prediction is not None:
    predicted_class = prediction[:,0]
    predicted_proba = prediction[:,1]
    
    df_mercado['classe_predita'] = predicted_class.astype(int)
    df_mercado['probabilidade'] = predicted_proba
    
    print('Accuracy of model on validation set: {:.3f}'
         .format(predictor.model.score(X, y)))
    
    df_mercado.pivot_table(index='passivel_homologacao',columns='classe_predita',values='descricao',aggfunc='count')

# mesmo que seja um item a ser preditor, o argumento deve ser uma lista
predictor.predict(['Repetidor De Sinal Internet 600mbps Branco 110v/220v Wifi'])

# para predizer a classe de vários itens, basta passar um iterável como parâmetro
to_predict = df_mercado.groupby('passivel_homologacao').sample(2,random_state=111)['descricao'].to_list()
predictor.predict(to_predict)

df_mercado.groupby('passivel_homologacao').sample(2,random_state=111)
