==> notebooks/27_meli_api_search_module_test.ipynb <==
# Preparação do ambiente

## Bibliotecas

from melisearch import MeliSearch

import pandas as pd
from pathlib import Path
from tqdm.auto import tqdm

# Desenvolvimento

ms = MeliSearch(output_folder=r'd:\temp\meli')

results = ms.search_item('iphone 15 pro max')

df = pd.DataFrame(results)
for id in tqdm(df.id.unique()):
    ms.search_item_details(id)

item_details_folder = Path(r'D:\Temp\meli\item_details')
item_details_files = [file for file in item_details_folder.glob('*.json')]
details = [ms.parse_item_details_file(file) for file in item_details_files]
df_details = pd.DataFrame(details)

df_details
